napi-snap7 v 1.2.3
